fix(codec)!: stop truncating scaled values, widen wide fields to f64
Three defects, found by fuzzing every layer that parses bus data.

A truncated frame aborted the decoder: seek() lets the cursor sit past the buffer, and the remaining-room computation underflowed. Any short frame addressed to a polymorphic PGN downed the node. Cursor arithmeticis now saturating.

Encoding truncated instead of rounding, shifting every scaled field one LSB towards zero: 23.767 at 0.001 came back as 23.766. 29 of the 42 default PGNs failed a decode/encode cycle.

f32 holds integers exactly only to 2^24, so a 32-bit latitude scaled by 1e-7 drifted up to 0.75 m. Fields wider than 24 bits now use f64.

BREAKING: 25 fields across 15 default-manifest PGNs change from f32 to f64; PGN 129025 grows from 8 to 16 bytes.

+++ tests/protocol/transport/fast_packet/integration_tests.rs
@@ -398,7 +398,7 @@ fn test_stress_100_pgns() {
         // Create an AIS message with varying data
         let mut ais = Pgn129040::new();
         ais.user_id = 1_000_000 + i;
-        ais.latitude = 45.0 + (i as f32 * 0.01);
+        ais.latitude = 45.0 + (i as f64 * 0.01);
 
         let mut buffer = [0u8; 64];
         let len = ais
